minor fixes

+++ dbas/query_helper.py
@@ -640,10 +640,12 @@ class QueryHelper(object):
 		:param issue:
 		:return: undefined
 		"""
-		db_user = DBDiscussionSession.query(User).filter_by(nickname=user).first()
-		logger('QueryHelper', 'del_history_of_user','user ' + str(db_user.uid))
-		DBDiscussionSession.query(History).filter_by(author_uid=db_user.uid).delete()
-		transaction.commit()
+		# maybe we are anonymous
+		if user:
+			db_user = DBDiscussionSession.query(User).filter_by(nickname=user).first()
+			logger('QueryHelper', 'del_history_of_user','user ' + str(db_user.uid))
+			DBDiscussionSession.query(History).filter_by(author_uid=db_user.uid).delete()
+			transaction.commit()
 
 	def sql_timestamp_pretty_print(self, ts, lang):
 		"""
